+++ audiolazy/lazy_synth.py
@@ -418,12 +418,10 @@ class TableLookupMeta(AbstractOperatorOverloaderMeta):
   TableLookup class, applying them to the table contents, elementwise.
   Table length and number of cycles should be equal for this to work.
   """
-  __operators__ = ("add radd sub rsub mul rmul pow rpow div rdiv mod rmod "
-                   "truediv rtruediv floordiv rfloordiv "
-                   "pos neg lshift rshift rlshift rrshift "
-                   "and rand or ror xor rxor invert ")
+  __operators__ = "+ - * / // % ** << >> & | ^ ~"
 
-  def __binary__(cls, op_func):
+  def __binary__(cls, op):
+    op_func = op.func
     def dunder(self, other):
       if isinstance(other, TableLookup):
         if self.cycles != other.cycles:
@@ -439,7 +437,8 @@ class TableLookupMeta(AbstractOperatorOverloaderMeta):
       raise NotImplementedError("Unknown action do be done")
     return dunder
 
-  def __rbinary__(cls, op_func):
+  def __rbinary__(cls, op):
+    op_func = op.func
     def dunder(self, other):
       if isinstance(other, (int, float, complex)):
         new_table = [op_func(other, data) for data in self.table]
@@ -447,7 +446,8 @@ class TableLookupMeta(AbstractOperatorOverloaderMeta):
       raise NotImplementedError("Unknown action do be done")
     return dunder
 
-  def __unary__(cls, op_func):
+  def __unary__(cls, op):
+    op_func = op.func
     def dunder(self):
       new_table = [op_func(data) for data in self.table]
       return TableLookup(new_table, self.cycles)